refactor(messages): extract message fetching logic into useMessages hook
This change improves code reusability and separation of concerns by moving the infinite query and pagination logic from MessagesView into a dedicated custom hook, reducing duplication and simplifying the component.
